scrapers/Crunchbase_API/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.116.2
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
multidict==6.6.4
mysql-connector-python==9.4.0
packaging==25.0
playwright==1.55.0
pluggy==1.6.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-dotenv==1.1.1
python-slugify==8.0.4
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
yarl==1.20.1

# Similarity search dependencies
numpy>=1.24.0
pandas>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
tqdm>=4.65.0
